add authorization for notifications and announcements

+++ src/main/java/pl/lodz/p/ias/io/powiadomienia/announcement/AnnouncementService.java
@@ -1,7 +1,11 @@
 package pl.lodz.p.ias.io.powiadomienia.announcement;
 
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import pl.lodz.p.ias.io.powiadomienia.notification.NotificationService;
+import pl.lodz.p.ias.io.powiadomienia.notification.NotificationType;
 import pl.lodz.p.ias.io.uwierzytelnianie.model.Account;
 import pl.lodz.p.ias.io.uwierzytelnianie.repositories.AccountRepository;
 
@@ -18,6 +22,7 @@ public class AnnouncementService {
     private AnnouncementRepository announcementRepository;
     private UserAnnouncementRepository userAnnouncementRepository;
     private AccountRepository accountRepository;
+    private NotificationService notificationService;
 
     /**
      * Tworzy nowe ogłoszenie i przypisuje je do każdego użytkownika.
@@ -37,17 +42,21 @@ public class AnnouncementService {
             userAnnouncementRepository.save(userAnnouncement);
         }
 
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        notificationService.notify("Successfully created Announcement.", NotificationType.INFORMATION, accountRepository.findByUsername(username));
+
         return savedAnnouncement;
     }
 
     /**
      * Pobiera listę ogłoszeń użytkownika.
      *
-     * @param userId ID użytkownika którego ogłoszenia zostaną pobrane
+     * @param username login użytkownika którego ogłoszenia zostaną pobrane
      * @return lista {@link Announcement} danego użytkownika
      */
-    public List<Announcement> getUsersAnnouncements(Long userId) {
-        List<UserAnnouncement> userAnnouncementList = userAnnouncementRepository.findAllByUser(accountRepository.findById(userId).get());
+    public List<Announcement> getUsersAnnouncements(String username) {
+        List<UserAnnouncement> userAnnouncementList = userAnnouncementRepository.findAllByUser(accountRepository.findByUsername(username));
         List<Announcement> announcementList = new ArrayList<>();
         for (UserAnnouncement userAnnouncement : userAnnouncementList) {
             announcementList.add(userAnnouncement.getAnnouncement());
@@ -62,6 +71,10 @@ public class AnnouncementService {
      * @param announcementId ID ogłoszenia do ukrycia
      */
     public void hideAnnouncement(Long announcementId) {
-        userAnnouncementRepository.deleteById(announcementId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        if(userAnnouncementRepository.findById(announcementId).get().getUser().getUsername().equals(username)) {
+            userAnnouncementRepository.deleteById(announcementId);
+        }
     }
 }
